requirements.txt
transformers
torch
accelerate
pillow
fastapi
uvicorn[standard]
python-multipart
requests
slowapi
prometheus-client
colorama